updated to supplier

+++ database/migrations/2023_05_12_090011_create_suppliers_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('address');
-            $table->unsignedBigInteger('contact_number');
+            $table->integer('contact_number');
             $table->timestamps();
         });
     }
